Add the feature of adding multiple passengers
Included the functionality of adding multiple passengers. User
can enter the details of a passenger and then either chose to enter another
passenger detail or to commit the transaction. Upon commiting the
transaction the details of the entered passenger/passengers and the
respective address/addresses will get stored in the database.